se-proxy: Add support for an interloper that hijacks conftool traffic.
The daemon now also listens on port 22223. When something's connected
there, traffic from it on 0xfffffffd (the conftool address) goes to the
most recently connected device, and traffic from a device to 0xfffffffd
goes to the most recent interloper. This allows manual commands to be
executed without disrupting telemetry to the portal.